feat: add middleware module and StdioTransport::layer()
Introduce a dedicated `src/middleware/` module for MCP-specific tower
layers and add `.layer()` to `StdioTransport` for transport parity
with `HttpTransport` and `WebSocketTransport`.

- Add `StdioTransport::layer()` that internally converts to
  `GenericStdioTransport` while preserving notification forwarding
- Create `middleware` module with `McpTracingLayer` (moved from
  `tracing_layer.rs`) and new `ToolCallLoggingLayer`
- `ToolCallLoggingLayer` intercepts only `CallTool` requests and emits
  structured tracing events with tool name, duration, and status
  (success/error/denied) at target `mcp::tools`
- Backward-compatible: `tracing_layer` module re-exports from
  `middleware`, crate root re-exports all middleware types
- Update `transport_middleware` example to demonstrate both features

Closes #488, closes #489
